Exercise 12 with .to_f

+++ ex12.rb
@@ -1,11 +1,11 @@
 print "Give me a number: "
-number = gets.chomp.to_i
+number = gets.chomp.to_f
 
 bigger = number * 100
 puts "A bigger number is #{bigger}."
 
 print "Give me another number: "
-another = gets.chomp.to_i
+another = gets.chomp.to_f
 
 smaller = number / 100
 puts "A smaller number is #{smaller}."
